Profile Stats Activity
The radar graph is on the Profile Stats page, as well as a spinner and
a progress bar. I have not implement a way to grab data and create a
chart from that data. If you guys can think of a way to do it, that will
be great. It's way late, my brain is fried and I've got a long drive
ahead of me tomorrow. So time to go night night.